add route security

+++ html/routes/articles/delete.php
@@ -2,20 +2,31 @@
 
 require_once __DIR__ . "/../../libraries/response.php";
 require_once __DIR__ . "/../../libraries/parameters.php";
+require_once __DIR__ . "/../../libraries/authorization.php";
 require_once __DIR__ . "/../../entities/articles/delete-articles.php";
 
-try {
-    $parameters = getParametersForRoute("/articles/:article");
-    $id = $parameters["article"];
-    deleteArticle($id);
 
-    echo jsonResponse(200, [], [
-        "success" => true,
-        "message" => "deleted"
-    ]);
-} catch (Exception $exception) {
-    echo jsonResponse(200, [], [
-        "success" => false,
-        "error" => $exception->getMessage()
-    ]);
+if (authorization(3)){
+    try {
+        $parameters = getParametersForRoute("/articles/:article");
+        $id = $parameters["article"];
+        deleteArticle($id);
+    
+        echo jsonResponse(200, [], [
+            "success" => true,
+            "message" => "deleted"
+        ]);
+    } catch (Exception $exception) {
+        echo jsonResponse(200, [], [
+            "success" => false,
+            "error" => $exception->getMessage()
+        ]);
+    }
 }
+
+
+    echo jsonResponse(400, [], [
+        "success" => false,
+        "error" => "Vous n'avez pas les droit néccessaire pour effectuer cette action"
+    ]);
+
